requirements.txt
pillow~=11.3.0
PyNaCl>=1.5.0
opencv-python~=4.12.0.88
pyotp~=2.9.0
qrcode~=8.0